vsc_repeat/lmms-eval/pyproject.toml
[tool.black]
line-length = 240

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lmms_eval"
version = "0.5.0"
authors = [
    { name = "LMMMs-Lab Evaluation Team", email = "lmms-lab@outlook.com" },
]
description = "A framework for evaluating large multi-modality language models"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
requires-python = ">=3.9"
dependencies = [
    "accelerate==0.34.2",
    "black==24.1.0",
    "isort==5.13.2",
    "datasets==2.16.1",
    "evaluate==0.4.3",
    "httpx==0.23.3",
    "jsonlines",
    "numexpr",
    "numpy==1.26.4",
    "peft==0.13.0",
    "pybind11==2.13.6",
    "pytablewriter",
    "sacrebleu==2.4.3",
    "scikit-learn==1.5.2",
    "sqlitedict==2.1.0",
    "torch==2.5.1",
    "torchvision==0.20.1",
    "timm",
    "einops",
    "ftfy",
    "openai",
    "opencv-python-headless",
    "av",
    "hf_transfer",
    "nltk",
    "sentencepiece==0.1.99",
    "yt-dlp",
    "pycocoevalcap",
    "tqdm-multiprocess",
    "transformers==4.39.2",
    "transformers-stream-generator",
    "zstandard",
    "pillow",
    "pyyaml",
    "sympy",
    "latex2sympy2",
    "mpmath",
    "Jinja2",
    "openpyxl",
    "loguru",
    "hf_transfer",
    "tenacity==8.3.0",
    "wandb==0.18.1",
    "tiktoken",
    "pre-commit",
    "pydantic",
    "packaging",
    "decord; platform_system != 'Darwin'",
    "eva-decord; platform_system == 'Darwin'",
    "zss",
    "protobuf==5.29.4",
    "sentence_transformers",
    "python-dotenv",
    # the following packages are for cambrians
    "ezcolorlog",
    "open_clip_torch==2.32.0",
    "diffusers==0.27.2",
    "huggingface-hub==0.25.2",
    "tensorflow==2.19.0",
]

[project.optional-dependencies]
audio = [
    "more-itertools",
    "editdistance",
    "zhconv",
    "librosa",
    "soundfile"
]
metrics = [
    "pywsd",
    "spacy",
    "anls",
    "rouge",
    "capture_metric",
    "Levenshtein",
]
gemini = [
    "google-generativeai",
]
reka = [
    "httpx==0.23.3",
    "reka-api",
]
qwen = [
    "decord",
    "qwen_vl_utils",
]
mmsearch = [
    "playwright",
    "requests",
    "matplotlib",
    "duckduckgo_search",
    "langchain",
    "langchain-community",
    "beautifulsoup4",
    "FlagEmbedding",
    "rouge",
]

[tool.setuptools.packages.find]
include = ["lmms_eval*"]
exclude = [
    "assets*",
    "benchmark*",
    "docs",
    "dist*",
    "playground*",
    "scripts*",
    "tests*",
    "checkpoints*",
    "project_checkpoints*",
    "debug_checkpoints*",
    "mlx_configs*",
    "wandb*",
    "notebooks*",
    "logs*",
]

[tool.wheel]
exclude = [
    "assets*",
    "benchmark*",
    "docs",
    "dist*",
    "playground*",
    "scripts*",
    "tests*",
    "checkpoints*",
    "project_checkpoints*",
    "debug_checkpoints*",
    "mlx_configs*",
    "wandb*",
    "notebooks*",
    "logs*",
]

[project.scripts]
lmms-eval = "lmms_eval.__main__:cli_evaluate"

[project.urls]
Homepage = "https://lmms-lab.github.io"
Repository = "https://github.com/EvolvingLMMs-Lab/lmms-eval"
